Add redirect_back w/ inertia errors support
This builds on the work in 40f36a5580a3362ed2e9dbd23373b61e82dfc98d
which added an inertia-aware wrapper around `redirect_to`.

This change implements the same inertia errors handling for
`redirect_back`, which I've found to be particularly useful when working
with Inertia, especially with error handling.